Formatting, Indexing, and Clarity
Java files have been autoformated for visibility. Added an indexing motor in both Shooter subsystem and command. RobotMap port variable names now include what type of port they are for - *_CAN for can devices, and *_SPARK for PWM devices (which, for this robot, is limited to sparks). The exception is the limit switch.